requirements.txt
amqp==5.0.6
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.18.34
botocore==1.21.34
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
Django==3.0.5
django-allauth==0.45.0
django-celery-results==2.2.0
django-cors-headers==3.8.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
dnspython==2.0.0
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
kombu==5.1.0
numpy==1.21.2
oauthlib==3.1.1
pandas==1.1.5
pika==1.2.0
Pillow==8.0.1
prettytable==2.2.0
prompt-toolkit==3.0.20
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==1.7.1
pymongo==3.11.2
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
stripe==2.56.0
tcping==0.1.1rc1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
